Cache: Ability to mock cache within one test
Mocks are registered using the builder which specify the return value
for different inputs.
Within PHPUnit it is common that the input must have the same order
as defined in the test which is not always given due to plugins etc.
We don't mind the order and just want to override values
so we build our own ProxyMocker which also delegates for undefined
inputs.

* Mock\Cache: Cache mock ability
* CacheAssertions: Shortcut for mocking cache
* doc: Updated readme with examples for cache mocks
* Minor code cleanupProxyInvocationBuilder