dev: Switch to Mockery
PHPUnit and Prophecy have the ability to mock objects and classes.
Unfortunately the PHPUnit Mock system changes very often
even within major versions so maintaining a compatibility
is very hard,
while Prophecy does not allow proxy calls to the original object.
Mockery solves both of that
and it is compatible with PHP 5.6 which is nice too.